// marketplace-ui/src/app/shared/enums/time-ago.enum.ts
export enum TimeAgo {
  YEAR_AGO = 'common.timeAgo.yearAgo',
  YEARS_AGO = 'common.timeAgo.yearsAgo',
  MONTH_AGO = 'common.timeAgo.monthAgo',
  MONTHS_AGO = 'common.timeAgo.monthsAgo',
  WEEK_AGO = 'common.timeAgo.weekAgo',
  WEEKS_AGO = 'common.timeAgo.weeksAgo',
  DAY_AGO = 'common.timeAgo.dayAgo',
  DAYS_AGO = 'common.timeAgo.daysAgo',
  HOUR_AGO = 'common.timeAgo.hourAgo',
  HOURS_AGO = 'common.timeAgo.hoursAgo',
  MINUTE_AGO = 'common.timeAgo.minuteAgo',
  MINUTES_AGO = 'common.timeAgo.minutesAgo',
  SECOND_AGO = 'common.timeAgo.secondAgo',
  SECONDS_AGO = 'common.timeAgo.secondsAgo',
}